app/requirements.txt
Flask==3.0.3
rembg==2.0.57
Pillow==10.3.0

